Move kernel to App namespace
