+++ bin/console
@@ -1,7 +1,7 @@
 #!/usr/bin/env php
 <?php
 
-use Enhavo\Kernel;
+use App\Kernel;
 use Symfony\Bundle\FrameworkBundle\Console\Application;
 use Symfony\Component\Console\Input\ArgvInput;
 use Symfony\Component\Debug\Debug;
