+++ src/App/CacheKernel.php
@@ -1,11 +1,6 @@
 <?php
-/**
- * Created by PhpStorm.
- * User: gseidel
- * Date: 2019-05-08
- */
 
-namespace Enhavo;
+namespace App;
 
 use Symfony\Bundle\FrameworkBundle\HttpCache\HttpCache;
 use Symfony\Component\HttpFoundation\Request;
